extract adapter to own class, proper encapsulation
