+++ app/src/main/java/com/tinnvec/dctvandroid/ChannelListCallback.java
@@ -1,0 +1,8 @@
+package com.tinnvec.dctvandroid;
+
+
+import com.tinnvec.dctvandroid.channel.AbstractChannel;
+
+public interface ChannelListCallback {
+    void onChannelClicked(AbstractChannel channel);
+}
